git-svn-id: file:///home/gecchi/repos@1669 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/src/jp/ggaf/dx9core/exception/GgafDx9CriticalException.cpp
@@ -10,7 +10,7 @@ string GgafDx9CriticalException::getMsg() {
     char buff[1000];
     ZeroMemory(buff, sizeof(buff));
 
-    sprintf(buff, "\nHRESULT=0x%08X(%d) is %s. ErrStr=%s Desc=%s", _hr, _hr, ((FAILED(_hr)) ? "FAILED" : "SUCCEEDED"),
+    sprintf(buff, "\nHRESULT=0x%08X(%d) is %s. ErrStr=%s Desc=%s", (UINT)_hr, (int)_hr, ((FAILED(_hr)) ? "FAILED" : "SUCCEEDED"),
             DXGetErrorString9(_hr), DXGetErrorDescription9(_hr));
     string hrstr = buff;
     return _message + hrstr;
